requirements.txt
python-dotenv==1.1.0
Telethon==1.40.0
pandas==2.2.3